+++ plugins/stage.lua
@@ -199,6 +199,18 @@ methods_stage = {
         list:clear()
     end,
 
+
+    get_room = function(self, variant)
+        local list = List.wrap(self.room_list)
+
+        if variant < 1 or variant > #list then
+            log.error("Variant must be between 1 and variant count (inclusive)", 2)
+            return
+        end
+
+        return list[variant]
+    end
+
 }